Refactoring Data Store Preferences
