+++ app/src/main/java/com/unsa/healthcare/data/database/preferences/DataStorePreferences.kt
@@ -1,4 +1,4 @@
-package com.unsa.healthcare.data.database
+package com.unsa.healthcare.data.database.preferences
 
 interface DataStorePreferences {
     suspend fun saveJwtToken(jwt: String)
