+++ app/src/main/java/com/unsa/healthcare/data/repositories/PreferencesRepository.kt
@@ -6,7 +6,7 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.preferencesDataStore
-import com.unsa.healthcare.data.database.DataStorePreferences
+import com.unsa.healthcare.data.database.preferences.DataStorePreferences
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
